Don't auto-submit a form if it doesn't have the post method
Some websites don't post their forms the 'normal' way but instead use
some js function triggered by the login button. This means they don't
have to set the form method to post, which leads to the login
credentials being posted visibly as get parameters of the current url
which we want to avoid.